+++ components/TokenLink/TokenLink.tsx
@@ -1,4 +1,4 @@
-﻿﻿import React from "react";
+﻿import React from "react";
 import {useRouter} from "next/router";
 import Link from "next/link";
